refactor: create execution/session/ subpackage from flat session modules
Move session.py → session/__init__.py and _session_model.py,
_session_content.py, _session_outcome.py, _retry_fsm.py into
execution/session/. Update all internal cross-references and
_headless_result.py's direct imports.
