+++ src/autoskillit/execution/session/_session_model.py
